Refactor: Improve page structure and component organization
This commit refactors the application by extracting page content into dedicated wrapper components, improving code organization and maintainability. It also includes several minor component and utility improvements.
