+++ new-nextjs-app/src/app/add-property/page.tsx
@@ -3,7 +3,7 @@
 import dynamic from "next/dynamic";
 
 // Import the AddProperty component directly instead of using dynamic import to avoid build issues
-import AddProperty from "@/app/properties/add/page";
+import AddProperty from "../properties/add/page";
 
 export default function AddPropertyShallow() {
   return <AddProperty />;
