+++ new-nextjs-app/src/app/page.tsx
@@ -1,10 +1,6 @@
 import React from "react";
 import type { Metadata } from 'next'
-import HeroSection from "@/components/home/HeroSection";
-import PropertiesSection from "@/components/home/PropertiesSection";
-import OwnerServiceBlock from "@/components/home/OwnerServiceBox";
-import ServiceBlocksGroup from "@/components/home/ServiceBlocksGroup";
-import Reviews from "@/components/common/footer/Reviews";
+import HomePageWrapper from "./HomePageWrapper";
 
 // SEO Metadata for Homepage
 export const metadata: Metadata = {
@@ -93,13 +89,7 @@ export default function HomePage() {
         }}
       />
       
-      <div className="font-poppins bg-[var(--color-bg-primary)] text-[var(--color-text-primary)]">
-        <HeroSection/>
-        <PropertiesSection />
-        <ServiceBlocksGroup />
-        <OwnerServiceBlock />
-        <Reviews />
-      </div>
+      <HomePageWrapper />
     </>
   );
 }
