+++ new-nextjs-app/src/components/home/BlurHeader.tsx
@@ -62,13 +62,13 @@ const BlurHeader = () => {
 
                   <div className="absolute left-1/2 -translate-x-1/2 top-full mt-2 w-48 opacity-0 invisible group-hover:opacity-100 group-hover:visible transition-all duration-300">
                     <div className="bg-[color:var(--color-surface)]/95 backdrop-blur-lg rounded-xl shadow-2xl border border-white/10 overflow-hidden">
-                      {item.items.map((subItem) => (
+                      {item.items.map((subItem, index) => (
                         <a
-                          key={subItem}
+                          key={index}
                           href="#"
                           className="block px-4 py-3 text-white/90 hover:text-white hover:bg-white/10 transition-colors duration-200 border-b border-white/5 last:border-0"
                         >
-                          {subItem}
+                          {typeof subItem === 'string' ? subItem : subItem.name}
                         </a>
                       ))}
                     </div>
@@ -133,13 +133,13 @@ const BlurHeader = () => {
                         className="overflow-hidden"
                       >
                         <div className="pl-4 py-2 space-y-2">
-                          {item.items.map((subItem) => (
+                          {item.items.map((subItem, index) => (
                             <a
-                              key={subItem}
+                              key={index}
                               href="#"
                               className="block px-4 py-2 text-white/80 hover:text-white hover:bg-[var(--color-primary)]/10 rounded-lg transition-colors duration-200"
                             >
-                              {subItem}
+                              {typeof subItem === 'string' ? subItem : subItem.name}
                             </a>
                           ))}
                         </div>
